Help Flags
Added help (`-h` and `--help`) flags to all coreutils that were missing them except `true` and `false` as well as updated all the man pages to fit the new functionality.

`#![deny(warings)]` was now reactivated as well for all altered files, as the unused `const` warning is no longer happening.